refactor: Replace stdint.h with custom type definitions and add clang-tidy support (#32)
* Added types.h

* Minor fixes of long

* Changed stddef to types.h

* Fixed minor missed stdint and stddef

* updated compile_commands

* Renamed clangd-tidy to clang-tidy

* Fixed some lints and changed sleep to ksleep

* Fixed linting issues

* Fixed compile errors on test and prod

* Format code

+++ src/arch/x86/gdt/gdt.c
@@ -1,12 +1,11 @@
 #include "arch/x86/gdt/gdt.h"
 #include "arch/x86/entry.h"
-
-#include <stdint.h>
-#include <string.h>
+#include "lib/string.h"
+#include "types.h"
 
 #define NUM_ENTRIES 6
 
-extern void gdt_flush(uint32_t);
+extern void gdt_flush(u32);
 extern void* stack_top;
 
 static tss_entry_t tss_entry;
@@ -15,8 +14,8 @@ static descriptor_pointer_t gdt_ptr;
 
 /* Private */
 
-static void gdt_set_gate(uint32_t num, uint32_t base, uint32_t limit,
-    uint8_t access, uint8_t gran)
+static void gdt_set_gate(u32 num, u32 base, u32 limit,
+    u8 access, u8 gran)
 {
     gdt_entries[num].lower_base = (base & 0xFFFF);
     gdt_entries[num].middle_base = (base >> 16) & 0xFF;
@@ -31,11 +30,11 @@ static void gdt_set_gate(uint32_t num, uint32_t base, uint32_t limit,
 
 static void tss_init(void)
 {
-    gdt_set_gate(5, (uint32_t)&tss_entry, sizeof(tss_entry_t) - 1, 0x89, 0x00);
+    gdt_set_gate(5, (u32)&tss_entry, sizeof(tss_entry_t) - 1, 0x89, 0x00);
     memset(&tss_entry, 0, sizeof(tss_entry_t));
 
     tss_entry.ss0 = 0x10;
-    tss_entry.esp0 = (uintptr_t)&stack_top;
+    tss_entry.esp0 = (u32)&stack_top;
 }
 
 /* Public */
@@ -43,7 +42,7 @@ static void tss_init(void)
 void gdt_init(void)
 {
     gdt_ptr.limit = sizeof(entry_t) * NUM_ENTRIES - 1;
-    gdt_ptr.base = (uint32_t)&gdt_entries;
+    gdt_ptr.base = (u32)&gdt_entries;
 
     gdt_set_gate(0, 0, 0, 0, 0);                // NULL Gate
     gdt_set_gate(1, 0, 0xFFFFFFFF, 0x9A, 0xCF); // Kernel Code Segment
@@ -53,6 +52,6 @@ void gdt_init(void)
 
     tss_init();
 
-    gdt_flush((uint32_t)&gdt_ptr);
-    __asm__ volatile("ltr %0" : : "r"((uint16_t)0x28));
+    gdt_flush((u32)&gdt_ptr);
+    __asm__ volatile("ltr %0" : : "r"((u16)0x28));
 }
